+++ c/include/libsbp/v4/bootload/MSG_BOOTLOADER_HANDSHAKE_RESP.h
@@ -100,7 +100,7 @@ int sbp_msg_bootloader_handshake_resp_version_strcmp(
  * @return Size of sbp_msg_bootloader_handshake_resp_t::version in wire
  * representation
  */
-uint8_t sbp_msg_bootloader_handshake_resp_version_encoded_len(
+size_t sbp_msg_bootloader_handshake_resp_version_encoded_len(
     const sbp_msg_bootloader_handshake_resp_t *msg);
 
 /**
@@ -113,7 +113,7 @@ uint8_t sbp_msg_bootloader_handshake_resp_version_encoded_len(
  * @param msg sbp_msg_bootloader_handshake_resp_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_bootloader_handshake_resp_version_space_remaining(
+size_t sbp_msg_bootloader_handshake_resp_version_space_remaining(
     const sbp_msg_bootloader_handshake_resp_t *msg);
 /**
  * Set sbp_msg_bootloader_handshake_resp_t::version
@@ -213,8 +213,8 @@ const char *sbp_msg_bootloader_handshake_resp_version_get(
  * @param msg sbp_msg_bootloader_handshake_resp_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_bootloader_handshake_resp_version_section_strlen(
-    const sbp_msg_bootloader_handshake_resp_t *msg, uint8_t section);
+size_t sbp_msg_bootloader_handshake_resp_version_section_strlen(
+    const sbp_msg_bootloader_handshake_resp_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_bootloader_handshake_resp_t
