+++ c/include/libsbp/v4/file_io/MSG_FILEIO_READ_DIR_REQ.h
@@ -107,7 +107,7 @@ int sbp_msg_fileio_read_dir_req_dirname_strcmp(
  * @param msg sbp_msg_fileio_read_dir_req_t instance
  * @return Size of sbp_msg_fileio_read_dir_req_t::dirname in wire representation
  */
-uint8_t sbp_msg_fileio_read_dir_req_dirname_encoded_len(
+size_t sbp_msg_fileio_read_dir_req_dirname_encoded_len(
     const sbp_msg_fileio_read_dir_req_t *msg);
 
 /**
@@ -120,7 +120,7 @@ uint8_t sbp_msg_fileio_read_dir_req_dirname_encoded_len(
  * @param msg sbp_msg_fileio_read_dir_req_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_fileio_read_dir_req_dirname_space_remaining(
+size_t sbp_msg_fileio_read_dir_req_dirname_space_remaining(
     const sbp_msg_fileio_read_dir_req_t *msg);
 /**
  * Set sbp_msg_fileio_read_dir_req_t::dirname
@@ -218,8 +218,8 @@ const char *sbp_msg_fileio_read_dir_req_dirname_get(
  * @param msg sbp_msg_fileio_read_dir_req_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_fileio_read_dir_req_dirname_section_strlen(
-    const sbp_msg_fileio_read_dir_req_t *msg, uint8_t section);
+size_t sbp_msg_fileio_read_dir_req_dirname_section_strlen(
+    const sbp_msg_fileio_read_dir_req_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_fileio_read_dir_req_t
