+++ c/include/libsbp/v4/linux/MSG_LINUX_PROCESS_FD_SUMMARY.h
@@ -101,7 +101,7 @@ int sbp_msg_linux_process_fd_summary_most_opened_strcmp(
  * @return Size of sbp_msg_linux_process_fd_summary_t::most_opened in wire
  * representation
  */
-uint8_t sbp_msg_linux_process_fd_summary_most_opened_encoded_len(
+size_t sbp_msg_linux_process_fd_summary_most_opened_encoded_len(
     const sbp_msg_linux_process_fd_summary_t *msg);
 
 /**
@@ -114,7 +114,7 @@ uint8_t sbp_msg_linux_process_fd_summary_most_opened_encoded_len(
  * @param msg sbp_msg_linux_process_fd_summary_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_linux_process_fd_summary_most_opened_space_remaining(
+size_t sbp_msg_linux_process_fd_summary_most_opened_space_remaining(
     const sbp_msg_linux_process_fd_summary_t *msg);
 /**
  * Return the number of sections in
@@ -123,7 +123,7 @@ uint8_t sbp_msg_linux_process_fd_summary_most_opened_space_remaining(
  * @param msg sbp_msg_linux_process_fd_summary_t instance
  * @return Number of sections in string
  */
-uint8_t sbp_msg_linux_process_fd_summary_most_opened_count_sections(
+size_t sbp_msg_linux_process_fd_summary_most_opened_count_sections(
     const sbp_msg_linux_process_fd_summary_t *msg);
 
 /**
@@ -249,7 +249,7 @@ bool sbp_msg_linux_process_fd_summary_most_opened_append_vprintf(
  * @return Pointer to C string, NULL on error
  */
 const char *sbp_msg_linux_process_fd_summary_most_opened_get_section(
-    const sbp_msg_linux_process_fd_summary_t *msg, uint8_t section);
+    const sbp_msg_linux_process_fd_summary_t *msg, size_t section);
 
 /**
  * Obtain the length of a section in
@@ -263,8 +263,8 @@ const char *sbp_msg_linux_process_fd_summary_most_opened_get_section(
  * @param section Section number
  * @return Length of section
  */
-uint8_t sbp_msg_linux_process_fd_summary_most_opened_section_strlen(
-    const sbp_msg_linux_process_fd_summary_t *msg, uint8_t section);
+size_t sbp_msg_linux_process_fd_summary_most_opened_section_strlen(
+    const sbp_msg_linux_process_fd_summary_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_linux_process_fd_summary_t
