+++ c/include/libsbp/v4/piksi/MSG_COMMAND_OUTPUT.h
@@ -95,7 +95,7 @@ int sbp_msg_command_output_line_strcmp(const sbp_msg_command_output_t *a,
  * @param msg sbp_msg_command_output_t instance
  * @return Size of sbp_msg_command_output_t::line in wire representation
  */
-uint8_t sbp_msg_command_output_line_encoded_len(
+size_t sbp_msg_command_output_line_encoded_len(
     const sbp_msg_command_output_t *msg);
 
 /**
@@ -108,7 +108,7 @@ uint8_t sbp_msg_command_output_line_encoded_len(
  * @param msg sbp_msg_command_output_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_command_output_line_space_remaining(
+size_t sbp_msg_command_output_line_space_remaining(
     const sbp_msg_command_output_t *msg);
 /**
  * Set sbp_msg_command_output_t::line
@@ -206,8 +206,8 @@ const char *sbp_msg_command_output_line_get(
  * @param msg sbp_msg_command_output_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_command_output_line_section_strlen(
-    const sbp_msg_command_output_t *msg, uint8_t section);
+size_t sbp_msg_command_output_line_section_strlen(
+    const sbp_msg_command_output_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_command_output_t
